fix: made lp related stuff perform better
