+++ src/AppProviders.tsx
@@ -13,8 +13,8 @@ import { ScrollToTop } from 'Routes/ScrollToTop'
 import { Zendesk } from 'components/Zendesk/Zendesk'
 import { AppProvider } from 'context/AppProvider/AppContext'
 import { BrowserRouterProvider } from 'context/BrowserRouterProvider/BrowserRouterProvider'
+import { FoxEthProvider } from 'context/FoxEthProvider/FoxEthProvider'
 import { I18nProvider } from 'context/I18nProvider/I18nProvider'
-import { LpHoldingsValueProvider } from 'context/LpHoldingsValueProvider/LpHoldingsValueProvider'
 import { ModalProvider } from 'context/ModalProvider/ModalProvider'
 import { PluginProvider } from 'context/PluginProvider/PluginProvider'
 import { TransactionsProvider } from 'context/TransactionsProvider/TransactionsProvider'
@@ -49,9 +49,9 @@ export function AppProviders({ children }: ProvidersProps) {
                       <ModalProvider>
                         <TransactionsProvider>
                           <AppProvider>
-                            <LpHoldingsValueProvider>
+                            <FoxEthProvider>
                               <DefiManagerProvider>{children}</DefiManagerProvider>
-                            </LpHoldingsValueProvider>
+                            </FoxEthProvider>
                           </AppProvider>
                         </TransactionsProvider>
                       </ModalProvider>
