+++ src/features/defi/providers/fox-eth-lp/components/FoxEthLpManager/Overview/FoxEthLpOverview.tsx
@@ -4,7 +4,7 @@ import { ethAssetId } from '@shapeshiftoss/caip'
 import { DefiModalContent } from 'features/defi/components/DefiModal/DefiModalContent'
 import { Overview } from 'features/defi/components/Overview/Overview'
 import { DefiAction } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
-import { useFoxEthLpBalances } from 'pages/Defi/hooks/useFoxEthLpBalances'
+import { useFoxEth } from 'context/FoxEthProvider/FoxEthProvider'
 import { useGetAssetDescriptionQuery } from 'state/slices/assetsSlice/assetsSlice'
 import { selectAssetById, selectSelectedLocale } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
@@ -12,7 +12,12 @@ import { useAppSelector } from 'state/store'
 import { foxAssetId, foxEthLpOpportunityName } from '../../../constants'
 
 export const FoxEthLpOverview = () => {
-  const { loading, opportunity, foxBalance, ethBalance } = useFoxEthLpBalances()
+  const {
+    foxEthLpOpportunity: opportunity,
+    lpFoxBalance: foxBalance,
+    lpEthBalance: ethBalance,
+    lpLoading: loading,
+  } = useFoxEth()
 
   const lpAsset = useAppSelector(state => selectAssetById(state, opportunity.assetId))
   const foxAsset = useAppSelector(state => selectAssetById(state, foxAssetId))
