+++ src/features/defi/providers/fox-eth-lp/components/FoxEthLpManager/Withdraw/components/Status.tsx
@@ -15,11 +15,11 @@ import { AssetIcon } from 'components/AssetIcon'
 import { StatusTextEnum } from 'components/RouteSteps/RouteSteps'
 import { Row } from 'components/Row/Row'
 import { RawText, Text } from 'components/Text'
+import { useFoxEth } from 'context/FoxEthProvider/FoxEthProvider'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { bnOrZero } from 'lib/bignumber/bignumber'
-import { useFoxEthLpBalances } from 'pages/Defi/hooks/useFoxEthLpBalances'
 import {
   selectAssetById,
   selectFirstAccountSpecifierByChainId,
@@ -38,7 +38,7 @@ export const Status = () => {
   const { state, dispatch } = useContext(WithdrawContext)
   const { query, history: browserHistory } = useBrowserRouter<DefiQueryParams, DefiParams>()
   const { chainId } = query
-  const { opportunity, getOpportunityData } = useFoxEthLpBalances()
+  const { foxEthLpOpportunity: opportunity } = useFoxEth()
   const {
     state: { wallet },
   } = useWallet()
@@ -79,9 +79,8 @@ export const Status = () => {
             : '0',
         },
       })
-      getOpportunityData()
     }
-  }, [confirmedTransaction, dispatch, ethAsset.precision, getOpportunityData])
+  }, [confirmedTransaction, dispatch, ethAsset.precision])
 
   const handleViewPosition = () => {
     browserHistory.push('/defi')
