+++ src/features/defi/providers/fox-farming/components/FoxFarmingManager/Deposit/FoxFarmingDeposit.tsx
@@ -14,11 +14,11 @@ import { useTranslate } from 'react-polyglot'
 import { useSelector } from 'react-redux'
 import { CircularProgress } from 'components/CircularProgress/CircularProgress'
 import { DefiStepProps, Steps } from 'components/DeFi/components/Steps'
+import { useFoxEth } from 'context/FoxEthProvider/FoxEthProvider'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { logger } from 'lib/logger'
-import { useFoxFarmingBalances } from 'pages/Defi/hooks/useFoxFarmingBalances'
 import {
   selectAssetById,
   selectMarketDataById,
@@ -50,11 +50,8 @@ export const FoxFarmingDeposit = () => {
   const assetId = toAssetId({ chainId, assetNamespace, assetReference })
   const asset = useAppSelector(state => selectAssetById(state, assetId))
   const marketData = useAppSelector(state => selectMarketDataById(state, assetId))
-  const { opportunities, loading: foxFarmingLoading } = useFoxFarmingBalances()
-  const opportunity = useMemo(
-    () => opportunities.find(e => e.contractAddress === contractAddress),
-    [contractAddress, opportunities],
-  )
+  const { foxFarmingOpportunities, farmingLoading: foxFarmingLoading } = useFoxEth()
+  const opportunity = foxFarmingOpportunities.find(e => e.contractAddress === contractAddress)
 
   // user info
   const chainAdapter = chainAdapterManager.get(chainId)
