+++ src/features/defi/providers/fox-farming/components/FoxFarmingManager/Deposit/components/Status.tsx
@@ -14,8 +14,6 @@ import { Row } from 'components/Row/Row'
 import { RawText, Text } from 'components/Text'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { bnOrZero } from 'lib/bignumber/bignumber'
-import { useFoxEthLpBalances } from 'pages/Defi/hooks/useFoxEthLpBalances'
-import { useFoxFarmingBalances } from 'pages/Defi/hooks/useFoxFarmingBalances'
 import {
   selectAssetById,
   selectFirstAccountSpecifierByChainId,
@@ -35,8 +33,6 @@ export const Status = () => {
   const history = useHistory()
   const { query, history: browserHistory } = useBrowserRouter<DefiQueryParams, DefiParams>()
   const { chainId, assetReference } = query
-  const { getOpportunitiesData: reloadFarmingOpportunities } = useFoxFarmingBalances()
-  const { getOpportunityData: reloadLpOpportunity } = useFoxEthLpBalances()
   const assetNamespace = 'erc20'
   const assetId = toAssetId({ chainId, assetNamespace, assetReference })
   const feeAssetId = toAssetId({
@@ -76,20 +72,10 @@ export const Status = () => {
             : '0',
         },
       })
-      if (confirmedTransaction.status === 'Confirmed') {
-        reloadFarmingOpportunities()
-        reloadLpOpportunity()
-      }
     }
-  }, [
-    confirmedTransaction,
-    dispatch,
-    feeAsset.precision,
-    reloadFarmingOpportunities,
-    reloadLpOpportunity,
-  ])
+  }, [confirmedTransaction, dispatch, feeAsset.precision])
 
-  if (!state) return null
+  if (!state || !dispatch || !opportunity) return null
 
   const { statusIcon, statusText, statusBg, statusBody } = (() => {
     switch (state.deposit.txStatus) {
