+++ src/features/defi/providers/fox-farming/components/FoxFarmingManager/Overview/Claim/ClaimStatus.tsx
@@ -14,8 +14,6 @@ import { SlideTransition } from 'components/SlideTransition'
 import { RawText } from 'components/Text'
 import { useBrowserRouter } from 'hooks/useBrowserRouter/useBrowserRouter'
 import { bnOrZero } from 'lib/bignumber/bignumber'
-import { useFoxEthLpBalances } from 'pages/Defi/hooks/useFoxEthLpBalances'
-import { useFoxFarmingBalances } from 'pages/Defi/hooks/useFoxFarmingBalances'
 import {
   selectAssetById,
   selectFirstAccountSpecifierByChainId,
@@ -84,9 +82,6 @@ export const ClaimStatus = () => {
   const feeAsset = useAppSelector(state => selectAssetById(state, feeAssetId))
   const feeMarketData = useAppSelector(state => selectMarketDataById(state, feeAssetId))
 
-  const { getOpportunitiesData: reloadFarmingOpportunities } = useFoxFarmingBalances()
-  const { getOpportunityData: reloadLpOpportunity } = useFoxEthLpBalances()
-
   const accountSpecifier = useAppSelector(state =>
     selectFirstAccountSpecifierByChainId(state, chainId),
   )
@@ -105,12 +100,8 @@ export const ClaimStatus = () => {
           ? bnOrZero(confirmedTransaction.fee.value).div(`1e${feeAsset.precision}`).toString()
           : '0',
       })
-      if (confirmedTransaction.status === 'Confirmed') {
-        reloadFarmingOpportunities()
-        reloadLpOpportunity()
-      }
     }
-  }, [confirmedTransaction, feeAsset.precision, reloadFarmingOpportunities, reloadLpOpportunity])
+  }, [confirmedTransaction, feeAsset.precision])
 
   return (
     <SlideTransition>
